Set up base for changing dialogue based on location
Created a current_location variable in the yarn script that gets updated externally when the player moves. Then rerouted the end of the conversation to a new navigation node, where it stalls by running a line of (invisible) dialogue, and then is advanced by an external script when ready to start conversation. Then it goes to an if-selection, which changes node selection based on current location.